Refactor ordersample to own file
